api/requirements.txt
chromadb==0.4.24
fastapi==0.110.0
langchain==0.1.13
langchain-community==0.0.29
langchain-openai==0.1.1
python-dotenv==1.0.1
uvicorn==0.29.0